associate_ai_engineer/open_ai_api/pyproject.toml
[project]
name = "open-ai-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.3.0",
    "python-dotenv>=1.1.1",
]
